refactor: merge and restructure runtime routing logic
