+++ test/mocks/i18n.options.ts
@@ -1,0 +1,1 @@
+export const localeCodes = []
